MEM-79: fix value diff highlight
the oldEp and newEp can have different frames. usually, when the
frame are different, we want to use the newEp. however, if the
func_name is the same, then we want to use the newEp BUT preserve
the unique_hash and frame_id so that the front will show the
proper value diff highlight.